[DOC] 칸반 / 노트 상세페이지 관련 액션 모듈 기존 note에서 분리하여 추가 #67
- 칸반 노트 인덱싱 기능 개발을 위해 기존의 노트 수정 액션을 위치(상태값) 변경 수정과 내용 변경 수정으로 나눌 예정
- 기존의 노트 관련 액션 모듈의 양이 많아져 칸반 페이지와 노트 상세페이지에서 일어나는 액션을 noteKanban 으로 분리함
- 각 페이지 및 컴포넌트 내 해당 모듈의 경로 변경
